Edit comments and move iters into lib file
